Renamed folders so that they now follow the convention of project-page rather feedback-page.
